+++ WDE.MapRenderer/Managers/MdxManager.cs
@@ -321,7 +321,7 @@ namespace WDE.MapRenderer.Managers
                 }
             }));
             
-            var md = new MeshData(vertices, normals, uv1, new uint[] { }, null, null, uv2, packedBones);
+            var md = new MeshData(vertices, normals, uv1, new ushort[] { }, null, null, uv2, packedBones);
             var mesh = meshManager.CreateMesh(md);
             mesh.SetSubmeshCount(skin.Batches.Length);
             // 1 : load items to define active geosets
@@ -537,7 +537,7 @@ namespace WDE.MapRenderer.Managers
                 if (isCharacterModel && !activeGeosets!.Contains(section.skinSectionId))
                     continue;
 
-                using var indices = new PooledArray<uint>(section.indexCount);
+                using var indices = new PooledArray<ushort>(section.indexCount);
                 for (int i = 0; i < Math.Min(section.indexCount, skin.Indices.Length - section.indexStart); ++i)
                     indices[i] = skin.Indices[section.indexStart + i];
                 
@@ -925,7 +925,7 @@ namespace WDE.MapRenderer.Managers
                 }
             }));
             
-            var md = new MeshData(vertices, normals, uv1, new uint[] { }, null, null, uv2, packedBones);
+            var md = new MeshData(vertices, normals, uv1, new ushort[] { }, null, null, uv2, packedBones);
             
             var mesh = meshManager.CreateMesh(md);
             mesh.SetSubmeshCount(skin.Batches.Length);
@@ -944,7 +944,7 @@ namespace WDE.MapRenderer.Managers
 
                 var section = skin.SubMeshes[batch.skinSectionIndex];
 
-                using var indices = new PooledArray<uint>(section.indexCount);
+                using var indices = new PooledArray<ushort>(section.indexCount);
                 for (int i = 0; i < Math.Min(section.indexCount, skin.Indices.Length - section.indexStart); ++i)
                 {
                     indices[i] = skin.Indices[section.indexStart + i];
@@ -1108,7 +1108,7 @@ namespace WDE.MapRenderer.Managers
                 }
             }));
             
-            var md = new MeshData(vertices, normals, uv1, new uint[] { }, null, null, uv2, packedBones);
+            var md = new MeshData(vertices, normals, uv1, new ushort[] { }, null, null, uv2, packedBones);
             
             var mesh = meshManager.CreateMesh(md);
             mesh.SetSubmeshCount(skin.Batches.Length);
@@ -1128,7 +1128,7 @@ namespace WDE.MapRenderer.Managers
                 var section = skin.SubMeshes[batch.skinSectionIndex];
                 
 
-                using var indices = new PooledArray<uint>(section.indexCount);
+                using var indices = new PooledArray<ushort>(section.indexCount);
                 for (int i = 0; i < Math.Min(section.indexCount, skin.Indices.Length - section.indexStart); ++i)
                 {
                     indices[i] = skin.Indices[section.indexStart + i];
@@ -1273,7 +1273,7 @@ namespace WDE.MapRenderer.Managers
                 }
             }));
             
-            var md = new MeshData(vertices, normals, uv1, new uint[] { }, null, null, uv2, packedBones);
+            var md = new MeshData(vertices, normals, uv1, new ushort[] { }, null, null, uv2, packedBones);
             
             var mesh = meshManager.CreateMesh(md);
             mesh.SetSubmeshCount(skin.Batches.Length);
@@ -1293,7 +1293,7 @@ namespace WDE.MapRenderer.Managers
                 var section = skin.SubMeshes[batch.skinSectionIndex];
                 
 
-                using var indices = new PooledArray<uint>(section.indexCount);
+                using var indices = new PooledArray<ushort>(section.indexCount);
                 for (int i = 0; i < Math.Min(section.indexCount, skin.Indices.Length - section.indexStart); ++i)
                 {
                     indices[i] = skin.Indices[section.indexStart + i];
